requirements.txt
torch
torchvision
numpy
matplotlib
pydantic
python-dotenv
geopy
pytest
folium
pandas
Pillow
fastapi
uvicorn[standard]
